fix lint errors

+++ wasm/src/app/project.rs
@@ -75,7 +75,7 @@ impl Project {
     }
 
     pub fn get_image(&self) -> RgbaImage {
-        ImageBuffer::from_fn(self.width as u32, self.height as u32, |x, y| {
+        ImageBuffer::from_fn(self.width, self.height, |x, y| {
             image::Rgba(self.get_compiled_pixel(x, y))
         })
     }
